Update system-settings-dialog-box-customization-tab.md
Just ran into this issue where I could not shut off admin mode (it caused issues on a client that upgraded to v9.)  The checkbox seems to be disabled in v8.2 as well.  The fix is to use the Power Platform Admin Center (where it can be edited).  Cheers, Nick