Update to the applies to section
On the doc there is an applies to Applies to Dynamics 365 for Customer Engagement apps version 9.x (on-premises).  Then in the requirements its says An Office 365 Enterprise E3 or E5 subscription is required to do Activity Logging.  Since there is no documentation that shows you can link the on-premise deployment with a O365 subscription these two lines contradict each other.  If there is a way to link these can that be added or the applies to section be modified to remove the on-premise mention.