Update set-up-booking-rules.md
Added "Booking rules are only available for the hourly view and not daily, weekly and monthly." based on the investigation on https://portal.microsofticm.com/imp/v3/incidents/details/177117362/home. There is a plan of having it working for other views but not for the moment.

I did try to add it as a note but I couldn't. 
>[!NOTE]
> Booking rules are only available for the hourly view and not daily, weekly and monthly.