Update types-of-fields.md
Adding additional info on how negative values appear in the Unified Interface.  This was the only place I could find where currency values are documented, I could not find the equivalent docs in power apps.  If there is docs for this in for Power Apps we also need to update that page.