BPA can only be imported for Dynamics 365 v9+
The Best Practice Analyzer downloads contains a solution file that applies only to Dynamics v9 or a higher version. Therefore, even if USD v4 is supported for Dynamics 8.2. It is not easily possible to apply the Best Practice Analyzer for this version via download. It is possible to change the version of the solution.xml file and then import again. This will work as a workaround.

The SolutionPackageVersion  value in the solution.xml file has to be changed from 9.0 to 8.2 or something else.