Correct TSI instructions
Final step for the Azure Time Series Insights connection is incorrect. The updateRecord method (https://docs.microsoft.com/en-us/power-apps/developer/model-driven-apps/clientapi/reference/xrm-webapi/updaterecord) must use the row GUID (msdyn_iotproviderinstanceid), not the name of the IoT Hub (msdyn_providerinstanceid). Of course you'll need to find the GUID - not sure if there is an easy out-of-the-box way to do so.