Update deployment-administrators.md
For questions reach out to Kenhub on teams. 
Support has seen several cases where the exception I added is the error the customer sees during an import and the resolution is the same as indicated in this 'known issue'. This exception should be exposed in this article so a customer can potentially find this and self service/resolve the issue without contacting support.