Update network-ports-for-microsoft-dynamics-365.md
Nowhere do we define the required value needed during setup when a custom port or custom instance name is used. This should be specified so that the customer can self service. Without specifying the server correctly under this condition the customer will get an exception such as "the network path was not found" or "the server was not found or was not accessible"