Update user-session-management.md
There is no reference on onprem to how custom session timeouts invalidates ADFS tokenlifetime values.

Online documentation on session management does have context around how this settings overrides ENTRA token timeouts.

I have checked this onprem and this occurs, so we should update docs. Its normal that ADFS dedicated teams do not allow applications to tell them to increase tokenlifetime values. As per this, the experience on CRM is unusable. Custom session timeouts are the solution here, but we need to document it on its relation to ADFS tokenlifetime values.